use Rcpp 0.12.0 generate RcppExports split the hashed_model_matrix to: - bswap_32 - hash_function - vector_converter
